// src/types/enums.ts
// src/types/enums.ts
export enum DateFilterOption {
    PAST_HOUR = "past_hour",
    PAST_3_HOURS = "past_3_hours",
    PAST_6_HOURS = "past_6_hours",
    PAST_12_HOURS = "past_12_hours",
    PAST_24_HOURS = "past_24_hours",
    PAST_7_DAYS = "past_7_days",
    PAST_30_DAYS = "past_30_days",
    PAST_90_DAYS = "past_90_days",
    PAST_180_DAYS = "past_180_days",
    CUSTOM = "custom",
    ALL_TIME = "all_time"
  }